# src/main/resources/openapi.yaml
openapi: 3.1.0
info:
  title: Product API
  description: Product API
  version: 1.0.0
servers:
  - url: http://localhost:8080
    description: Local server
paths:
    /products:
        get:
          summary: List of products
          operationId: listProducts
          tags:
            - Product
          responses:
              '200':
                description: OK
                content:
                  application/json:
                    schema:
                      $ref: '#/components/schemas/ProductList'
        post:
          summary: Create Product
          operationId: createProduct
          tags:
            - Product
          requestBody:
            description: Optional description in *Markdown*
            required: true
            content:
              application/json:
                schema:
                  $ref: "#/components/schemas/ProductV1Request"
          responses:
            201:
              description: OK
              content:
                application/json:
                  schema:
                    $ref: "#/components/schemas/ProductV1"
            500:
              description: Failed to create
    /products/{id}:
        get:
          summary: Get a product by id
          operationId: getProduct
          tags:
            - Product
          parameters:
            - name: id
              in: path
              required: true
              schema:
                type: string
                format: uuid
          responses:
              '200':
                description: OK
                content:
                  application/json:
                    schema:
                      $ref: '#/components/schemas/ProductV1'
              '404':
                description: Not found
        delete:
          summary: Delete a product by id
          operationId: deleteProduct
          tags:
            - Product
          parameters:
            - name: id
              in: path
              required: true
              schema:
                type: string
                format: uuid
          responses:
              '200':
                description: OK
              '404':
                description: Not found

components:
  schemas:
    ProductList:
        type: array
        items:
            $ref: '#/components/schemas/ProductV1'
    ProductV1:
      type: object
      required:
        - id
        - name
        - description
        - priceInCents
        - createdAt
        - updatedAt
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        description:
           type: string
        priceInCents:
          type: integer
          format: int64
        createdAt:
          type: string
        updatedAt:
          type: string
    ProductV1Request:
      type: object
      required:
        - name
        - priceInCents
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        description:
          type: string
        priceInCents:
          type: integer
          format: int64